# Speed up debug builds