--- a/benchmarks/Cargo.toml
+++ b/benchmarks/Cargo.toml
@@ -7,7 +7,7 @@ publish = false
 [dependencies]
 anyhow = "1.0.56"
 csv = "1.1.6"
-milli = { path = "../milli" }
+milli = { path = "../milli", default-features = false }
 mimalloc = { version = "0.1.29", default-features = false }
 serde_json = { version = "1.0.79", features = ["preserve_order"] }
 
@@ -23,6 +23,9 @@ bytes = "1.1.0"
 convert_case = "0.5.0"
 flate2 = "1.0.22"
 reqwest = { version = "0.11.9", features = ["blocking", "rustls-tls"], default-features = false }
+
+[features]
+default = ["milli/default"]
 
 [[bench]]
 name = "search_songs"