--- a/cli/Cargo.toml
+++ b/cli/Cargo.toml
@@ -12,9 +12,12 @@ color-eyre = "0.6.1"
 csv = "1.1.6"
 eyre = "0.6.7"
 indicatif = "0.16.2"
-milli = { path = "../milli" }
+milli = { path = "../milli", default-features = false }
 mimalloc = { version = "0.1.29", default-features = false }
 serde = "1.0.136"
 serde_json = "1.0.79"
 stderrlog = "0.5.1"
 structopt = "0.3.26"
+
+[features]
+default = ["milli/default"]