@Kerollmops 2022-08-17
Nice!

```
LLVM ERROR: IO failure on output stream: no space on device
error: could not compile `lindera-ipadic`
```

bors merge