@Kerollmops 2022-08-17
@loiclec Is this PR making the library bigger or what? 😂