@loiclec 2022-08-17
I was starting to wonder as well. Since it is compiled with `opt-level = 0`, the binary will be bigger indeed :/ I'll check on my computer if it seems excessive or not.