@loiclec 2022-08-18
On my computer (macOS, so not necessarily the same), there doesn't seem to be any significant difference between the size of the `target/` folder before and after the PR. So I am still confused as to why it fails :(
Let's hope the CI was just having a bad day?

bors merge pretty please 🥺🙏