@loiclec 2022-08-18
nooooo 😭 I guess there really is something about that PR that makes rustc on Windows sad