@loiclec 2022-10-12
I've rebased the branch on main so it's ready to merge again :-)