@ManyTheFish 2022-10-12
Hey @loiclec, could it be possible to opt-in to the default features of Charabia when it is compiled in debug?
Most of the time, we don't need to have full language support activated during the development phase.